+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemShulkerBoxStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemStuffStorableComponent;
 
-public interface ItemShulkerBoxStack extends ItemStack {
+public interface ItemShulkerBoxStack extends ItemStack, ItemStuffStorableComponent {
 }