+++ server/src/main/java/org/allaymc/server/item/impl/ItemShulkerBoxStackImpl.java
@@ -2,7 +2,7 @@ package org.allaymc.server.item.impl;
 
 import lombok.experimental.Delegate;
 import org.allaymc.api.component.interfaces.Component;
-import org.allaymc.api.item.component.ItemItemStorableComponent;
+import org.allaymc.api.item.component.ItemStuffStorableComponent;
 import org.allaymc.api.item.initinfo.ItemStackInitInfo;
 import org.allaymc.api.item.interfaces.ItemShulkerBoxStack;
 import org.allaymc.server.component.interfaces.ComponentProvider;
@@ -11,7 +11,7 @@ import java.util.List;
 
 public class ItemShulkerBoxStackImpl extends ItemStackImpl implements ItemShulkerBoxStack {
     @Delegate
-    protected ItemItemStorableComponent itemStorableComponent;
+    protected ItemStuffStorableComponent stuffStorableComponent;
 
     public ItemShulkerBoxStackImpl(ItemStackInitInfo initInfo, List<ComponentProvider<? extends Component>> componentProviders) {
         super(initInfo, componentProviders);